move retry behavior into client with callback to generate streaming body
